# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes",
        "contact": {
            "email": "admin@blossombuddy.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout a user",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/plant": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get a list of plants",
                "operationId": "d5241537f6505ae4a717f0c92bbb40fb",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Create a new plant",
                "operationId": "8309c05e5772465302dd924140b3f526",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "common_name",
                                    "watering_general_benchmark"
                                ],
                                "properties": {
                                    "common_name": {
                                        "type": "string"
                                    },
                                    "watering_general_benchmark": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/plant/{name}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get a specific plant by name",
                "operationId": "80c395237d1c7867d3775bc0bdbb3736",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            }
        },
        "/plant/{id}": {
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Delete a specific plant",
                "operationId": "b1c2b45d495a4aadcd62ac01e6e7bcf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            }
        },
        "/user/plants": {
            "get": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Get a list of plants for the authenticated user",
                "operationId": "0eb9a070f9e58bc55f14fa36d1fa5031",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/user/plant": {
            "post": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Create a new plant for the authenticated user",
                "operationId": "b4273403948f66e51d9e0bbc5532127b",
                "parameters": [
                    {
                        "name": "common_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "watering_general_benchmark",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Plant created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/user/plant/{id}": {
            "delete": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Delete a plant for the authenticated user",
                "operationId": "b8bc9546b7a4d8703ee245e3fd461fe7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Plants",
            "description": "Plants"
        },
        {
            "name": "User Plants",
            "description": "User Plants"
        }
    ]
}